+++ workbench/src/renderer/menubar/report.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
-import ReactDom from 'react-dom';
-import {createRoot} from 'react-dom/client';
+import { createRoot } from 'react-dom/client';
 import { Translation } from 'react-i18next';
 
 import i18n from '../i18n/i18n';