+++ workbench/tests/renderer/plugin.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { ipcRenderer } from 'electron';
 import '@testing-library/jest-dom';
 import {
-  within, render, waitFor,
+  within, render, waitFor, fireEvent, createEvent,
 } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
@@ -17,7 +17,16 @@ import App from '../../src/renderer/app';
 
 jest.mock('../../src/renderer/server_requests');
 
-describe('Add plugin modal', () => {
+const PLUGIN_SETTING_ITEM = {
+  foo: {
+    modelTitle: 'Foo',
+    type: 'plugin',
+    version: '1.0',
+    env: 'foo/bar/baz/',
+  },
+};
+
+describe('Manage Plugins modal', () => {
   beforeEach(() => {
     getSpec.mockResolvedValue({
       model_id: 'foo',
@@ -53,12 +62,7 @@ describe('Add plugin modal', () => {
       spy = ipcRenderer.invoke.mockImplementation((channel, setting) => {
         if (channel === ipcMainChannels.GET_SETTING) {
           if (setting === 'plugins') {
-            return Promise.resolve({
-              foo: {
-                modelTitle: 'Foo',
-                type: 'plugin',
-              },
-            });
+            return Promise.resolve(PLUGIN_SETTING_ITEM);
           }
         } else if (channel === ipcMainChannels.HAS_MSVC) {
           return Promise.resolve(true);
@@ -140,12 +144,7 @@ describe('Add plugin modal', () => {
     const spy = ipcRenderer.invoke.mockImplementation((channel, setting) => {
       if (channel === ipcMainChannels.GET_SETTING) {
         if (setting === 'plugins') {
-          return Promise.resolve({
-            foo: {
-              modelTitle: 'Foo',
-              type: 'plugin',
-            },
-          });
+          return Promise.resolve(PLUGIN_SETTING_ITEM);
         }
       } else if (channel === ipcMainChannels.HAS_MSVC) {
         return Promise.resolve(true);
@@ -186,7 +185,7 @@ describe('Add plugin modal', () => {
 
   test('Add a plugin: failure with error displayed', async () => {
     const errorString = 'Failed to clone repository.';
-    const spy = ipcRenderer.invoke.mockImplementation((channel) => {
+    const spy = ipcRenderer.invoke.mockImplementation((channel, setting) => {
       if (channel === ipcMainChannels.HAS_MSVC) {
         return Promise.resolve(true);
       }
@@ -194,6 +193,11 @@ describe('Add plugin modal', () => {
         return Promise.reject(
           new Error(errorString)
         );
+      }
+      if (channel === ipcMainChannels.GET_SETTING) {
+        if (setting === 'plugins') {
+          return Promise.resolve({});
+        }
       }
       return Promise.resolve();
     });
@@ -224,15 +228,10 @@ describe('Add plugin modal', () => {
     ipcRenderer.invoke.mockImplementation((channel, setting) => {
       if (channel === ipcMainChannels.GET_SETTING) {
         if (setting === 'plugins') {
-          return Promise.resolve({
-            foo: {
-              modelTitle: 'Foo',
-              type: 'plugin',
-            },
-          });
+          return Promise.resolve(PLUGIN_SETTING_ITEM);
         }
       } else if (channel === ipcMainChannels.LAUNCH_PLUGIN_SERVER) {
-        return 1;
+        return 1111; // a fake PID
       }
       return Promise.resolve();
     });
@@ -257,22 +256,20 @@ describe('Add plugin modal', () => {
   });
 
   test('Remove a plugin', async () => {
-    let plugins = {
-      foo: {
-        modelTitle: 'Foo',
-        type: 'plugin',
-        version: '1.0',
-      },
-    };
+    let plugins = PLUGIN_SETTING_ITEM;
     const spy = ipcRenderer.invoke.mockImplementation((channel, setting) => {
       if (channel === ipcMainChannels.GET_SETTING) {
         if (setting === 'plugins') {
           return Promise.resolve(plugins);
         }
       } else if (channel === ipcMainChannels.REMOVE_PLUGIN) {
+        // after REMOVE_PLUGIN there will be a subsequent call to GET_SETTING,
+        // so this effectively replaces the mocked settings data
         plugins = {};
       } else if (channel === ipcMainChannels.HAS_MSVC) {
         return Promise.resolve(true);
+      } else if (channel === ipcMainChannels.LAUNCH_PLUGIN_SERVER) {
+        return 1111; // a fake PID
       }
       return Promise.resolve();
     });
@@ -299,5 +296,293 @@ describe('Add plugin modal', () => {
     // expect the plugin to have disappeared from the model list and the dropdown
     await waitFor(() => expect(queryByRole('button', { name: /Foo/ })).toBeNull());
     await waitFor(() => expect(queryByRole('option', { name: /Foo/ })).toBeNull());
+  });
+
+  test('Change the conda executable', async () => {
+    ipcRenderer.invoke.mockImplementation((channel, setting) => {
+      if (channel === ipcMainChannels.GET_SETTING) {
+        if (setting === 'plugins') {
+          return Promise.resolve({});
+        } else if (setting === 'micromamba') {
+          return Promise.resolve('micromamba')
+        }
+      } else if (channel === ipcMainChannels.SHOW_OPEN_DIALOG) {
+        return Promise.resolve({ filePaths: ['foo'] })
+      } else if (channel === ipcMainChannels.HAS_MSVC) {
+        return Promise.resolve(true);
+      }
+      return Promise.resolve();
+    });
+    const {
+      findByText, findByRole, findByLabelText,
+    } = render(<App />);
+
+    const spy = jest.spyOn(ipcRenderer, 'send');
+    await userEvent.click(await findByRole('button', { name: 'menu' }));
+    const managePluginsButton = await findByText(/Manage plugins/i);
+    await userEvent.click(managePluginsButton);
+
+    const input = await findByLabelText('Conda or mamba executable');
+    await userEvent.clear(input);
+    await userEvent.type(input, 'conda');
+    await waitFor(() => expect(input).toHaveValue('conda'));
+
+    await userEvent.click(await findByRole(
+      'button', { name: /browse for conda executable/ }));
+    await waitFor(() => { expect(input).toHaveValue('foo'); });
+
+    const file = new File([], 'my-conda');
+    fireEvent.dragEnter(input, {dataTransfer: {files: [file]}});
+    expect(input).toHaveClass('input-dragging');
+    fireEvent.drop(input, {dataTransfer: {files: [file]}});
+  
+    await waitFor(() => {
+      expect(input).not.toHaveClass('input-dragging');
+      expect(input).toHaveValue('my-conda');
+      expect(input).toHaveFocus();
+    });
+
+    const div = await findByLabelText('Configure conda executable (Advanced)')
+    const saveButton = await within(div).findByRole('button', { name: /Save/ });
+    await userEvent.click(saveButton);
+    await waitFor(() => {
+      expect(spy).toHaveBeenCalledWith(
+        ipcMainChannels.SET_SETTING,
+        'userDefinedMicromamba',
+        'my-conda'
+      );
+    });
+
+    const resetButton = await within(div).findByRole('button', { name: /Reset/ });
+    await userEvent.click(resetButton);
+    await waitFor(() => { expect(input).toHaveValue('micromamba'); });
+  });
+
+  test('Change a plugin env', async () => {
+    ipcRenderer.invoke.mockImplementation((channel, setting) => {
+      if (channel === ipcMainChannels.GET_SETTING) {
+        if (setting === 'plugins') {
+          return Promise.resolve(PLUGIN_SETTING_ITEM);
+        } else if (setting === 'plugins.foo.env') {
+          return Promise.resolve(PLUGIN_SETTING_ITEM.foo.env);
+        }
+      } else if (channel === ipcMainChannels.SHOW_OPEN_DIALOG) {
+        return Promise.resolve({ filePaths: ['/path/to/my_env'] });
+      } else if (channel === ipcMainChannels.HAS_MSVC) {
+        return Promise.resolve(true);
+      }
+      return Promise.resolve();
+    });
+    const {
+      findByText, findByRole, findByLabelText
+    } = render(<App />);
+
+    const spy = jest.spyOn(ipcRenderer, 'send');
+    await userEvent.click(await findByRole('button', { name: 'menu' }));
+    const managePluginsButton = await findByText(/Manage plugins/i);
+    await userEvent.click(managePluginsButton);
+
+    const div = await findByLabelText('Configure plugin environments (Advanced)')
+    const input = await within(div).findByLabelText('foo');
+    expect(input).toHaveValue(PLUGIN_SETTING_ITEM.foo.env);
+    await userEvent.clear(input);
+    await userEvent.type(input, 'my_env');
+    await waitFor(() => expect(input).toHaveValue('my_env'));
+
+    await userEvent.click(await findByRole(
+      'button', { name: /browse for env/ }));
+    await waitFor(() => { expect(input).toHaveValue('/path/to/my_env'); });
+
+    const saveButton = await within(div).findByRole('button', { name: /Save/ });
+    await userEvent.click(saveButton);
+    await waitFor(() => {
+      expect(spy).toHaveBeenCalledWith(
+        ipcMainChannels.SET_SETTING,
+        'plugins.foo.userDefinedEnv',
+        '/path/to/my_env'
+      );
+    });
+
+    const resetButton = await within(div).findByRole('button', { name: /Reset/ });
+    await userEvent.click(resetButton);
+    await waitFor(() => { expect(input).toHaveValue(PLUGIN_SETTING_ITEM.foo.env); });
+  });
+
+  test('Drag-and-drop populates the local plugin path', async () => {
+    ipcRenderer.invoke.mockImplementation((channel, setting) => {
+      if (channel === ipcMainChannels.GET_SETTING) {
+        if (setting === 'plugins') {
+          return Promise.resolve({});
+        }
+      } else if (channel === ipcMainChannels.HAS_MSVC) {
+        return Promise.resolve(true);
+      }
+  
+      return Promise.resolve();
+    });
+  
+    const {
+      findByText, findByRole, findByLabelText,
+    } = render(<App />);
+  
+    await userEvent.click(await findByRole('button', { name: 'menu' }));
+    await userEvent.click(await findByText(/Manage plugins/i));
+  
+    await userEvent.selectOptions(
+      await findByLabelText('Install from'),
+      'local path'
+    );
+  
+    const input = await findByLabelText('Local absolute path');
+    const file = new File([], 'plugin-dir');
+  
+    fireEvent.dragEnter(input, {
+      dataTransfer: { files: [file] },
+    });
+  
+    expect(input).toHaveClass('input-dragging');
+  
+    fireEvent.drop(input, {
+      dataTransfer: { files: [file] },
+    });
+  
+    await waitFor(() => {
+      expect(input).not.toHaveClass('input-dragging');
+      expect(input).toHaveValue('plugin-dir');
+      expect(input).toHaveFocus();
+    });
+  });
+
+  test('Drag-and-drop populates a plugin environment input', async () => {
+    ipcRenderer.invoke.mockImplementation((channel, setting) => {
+      if (channel === ipcMainChannels.GET_SETTING) {
+        if (setting === 'plugins') {
+          return Promise.resolve(PLUGIN_SETTING_ITEM);
+        }
+        if (setting === 'micromamba') {
+          return Promise.resolve('micromamba');
+        }
+      } else if (channel === ipcMainChannels.HAS_MSVC) {
+        return Promise.resolve(true);
+      }
+  
+      return Promise.resolve();
+    });
+  
+    const {
+      findByText, findByRole, findByLabelText,
+    } = render(<App />);
+  
+    await userEvent.click(await findByRole('button', { name: 'menu' }));
+    await userEvent.click(await findByText(/Manage plugins/i));
+  
+    const div = await findByLabelText(
+      'Configure plugin environments (Advanced)'
+    );
+    const input = await within(div).findByLabelText('foo');
+  
+    const file = new File([], 'plugin-environment');
+  
+    fireEvent.dragEnter(input, {
+      dataTransfer: {
+        files: [file],
+      },
+    });
+  
+    expect(input).toHaveClass('input-dragging');
+  
+    fireEvent.drop(input, {
+      dataTransfer: {
+        files: [file],
+      },
+    });
+  
+    await waitFor(() => {
+      expect(input).not.toHaveClass('input-dragging');
+      expect(input).toHaveValue('plugin-environment');
+      expect(input).toHaveFocus();
+    });
+  });
+
+  test('Drag-leave removes input-dragging from a filepath input', async () => {
+    ipcRenderer.invoke.mockImplementation((channel, setting) => {
+      if (channel === ipcMainChannels.GET_SETTING) {
+        if (setting === 'plugins') {
+          return Promise.resolve(PLUGIN_SETTING_ITEM);
+        }
+      } else if (channel === ipcMainChannels.HAS_MSVC) {
+        return Promise.resolve(true);
+      }
+  
+      return Promise.resolve();
+    });
+  
+    const {
+      findByText, findByRole, findByLabelText,
+    } = render(<App />);
+  
+    await userEvent.click(await findByRole('button', { name: 'menu' }));
+    await userEvent.click(await findByText(/Manage plugins/i));
+  
+    await userEvent.selectOptions(
+      await findByLabelText('Install from'),
+      'path'
+    );
+  
+    const input = await findByLabelText('Local absolute path');
+    const file = new File([], 'plugin-directory');
+  
+    fireEvent.dragEnter(input, {
+      dataTransfer: {
+        files: [file],
+      },
+    });
+  
+    expect(input).toHaveClass('input-dragging');
+  
+    fireEvent.dragLeave(input, {
+      dataTransfer: {
+        files: [file],
+      },
+    });
+  
+    expect(input).not.toHaveClass('input-dragging');
+  });
+
+  test('Drag-and-drop on the git URL input does nothing', async () => {
+    ipcRenderer.invoke.mockImplementation((channel, setting) => {
+      if (channel === ipcMainChannels.GET_SETTING) {
+        if (setting === 'plugins') {
+          return Promise.resolve({});
+        }
+      } else if (channel === ipcMainChannels.HAS_MSVC) {
+        return Promise.resolve(true);
+      }
+  
+      return Promise.resolve();
+    });
+  
+    const {
+      findByText, findByRole, findByLabelText,
+    } = render(<App />);
+  
+    await userEvent.click(await findByRole('button', { name: 'menu' }));
+    await userEvent.click(await findByText(/Manage plugins/i));
+  
+    const input = await findByLabelText('Git URL');
+  
+    const file = new File([], 'plugin-directory');
+    const dropEvent = createEvent.drop(input, {
+      dataTransfer: {
+        files: [file],
+      },
+    });
+    
+    const preventDefaultSpy = jest.spyOn(dropEvent, 'preventDefault');
+    
+    fireEvent(input, dropEvent);
+    
+    expect(preventDefaultSpy).toHaveBeenCalled();
+    expect(input).toHaveValue('');
   });
 });